chore: move chain utility functions from CLI to config crate
