+++ crates/script/src/broadcast.rs
@@ -15,7 +15,7 @@ use alloy_serde::WithOtherFields;
 use eyre::{Context, Result, bail};
 use forge_verify::provider::VerificationProviderType;
 use foundry_cheatcodes::Wallets;
-use foundry_cli::utils::{has_batch_support, has_different_gas_calc};
+use foundry_config::{has_batch_support, has_different_gas_calc};
 use foundry_common::{
     TransactionMaybeSigned,
     provider::{RetryProvider, get_http_provider, try_get_http_provider},
