+++ crates/script/src/simulate.rs
@@ -16,7 +16,8 @@ use dialoguer::Confirm;
 use eyre::{Context, Result};
 use forge_script_sequence::{ScriptSequence, TransactionWithMetadata};
 use foundry_cheatcodes::Wallets;
-use foundry_cli::utils::{has_different_gas_calc, now};
+use foundry_cli::utils::now;
+use foundry_config::has_different_gas_calc;
 use foundry_common::{ContractData, shell};
 use foundry_evm::traces::{decode_trace_arena, render_trace_arena};
 use futures::future::{join_all, try_join_all};
